ml-service/requirements.txt
fastapi==0.127.1
uvicorn==0.29.0
gunicorn==23.0.0

pydantic==2.12.5
python-dotenv==1.2.1

numpy==2.4.0
pandas==2.3.3
scikit-learn==1.8.0
joblib==1.5.3
xgboost==3.1.2
